+++ app/class-core.php
@@ -178,6 +178,7 @@ class Core {
 	 * @see Integrations\Spectra
 	 * @see Integrations\Wpml
 	 * @see Integrations\Shortpixel
+	 * @see Integrations\JS_Composer
 	 */
 	private function init_integrations() {
 		$loader = Loader::get_instance();
@@ -189,6 +190,7 @@ class Core {
 		$loader->integration( 'wpml' );
 		$loader->integration( 'shortpixel' );
 		$loader->integration( 'elementor' );
+		$loader->integration( 'js-composer' );
 	}
 
 	/**